Added drinks at Créa.
I forgot to indicate that the last two hours (3 to 5pm) were "booked" to have drinks and chat in an informal way at CREA café. After 3pm, we don't have the room anymore (although we can dispose of another one). CREA Café is located on the same campus and is very "gezellig" (=cosy, welcoming, informal).